requirements.dev.txt
pylint
python-dotenv
flask
flask_cors
flask_jwt_extended
pymongo
datetime